Add files via upload
Group project - DATA SCIENCE 2021
How does a country's economic policies affect personal freedom?
We developed a Machine Learning algorithm to predict if a country has a Personal Freedom score that can be classed as critical, based on the features used to measure its Economic Freedom. This code is split in two parts: an exploratory analysis and a predictive analysis. The dataset used is the Human Freedom Index 2019: A Global
Measurement of Personal, Civil & Economic Freedom, taken by the Cato institute from Kaggle.